Store uploaded image links in Mongo
Fix cloudinary configuration to create folder when we upload images. Adjust
campground model images to take in an array of images. When we upload an
image to cloudinary we store that url(path) to that image and we can access
that url using the req.files property in the campgrounds controller. We use
that url to display our image