Set up campgrounds edit page for image deletion
In our campgrounds edit template we displayed all of the images that
belong to that campground and added checkboxes next to them. Each checkbox
will store the path of that image it is correlated to as its value.
It will also have an array which will store the values(image paths) of
each image that we wish to delete.