Display map using Mapbox
Imported scripts to display map from Mapbox. Quick note: hardware
acceleration must be on in google chrome for map to work. Fixed issue
with seeds. Was getting a validation error when trying to re-seed database.
Error stated that geometry.type is required. Added another property in the
seeds index file to resolve issue.